+++ CHANGES.rst
@@ -3,6 +3,7 @@ Changelog
 
 
 - Add ``HashFS.size()`` method that returns the size of all files in bytes.
+- Add ``HashFS.count()``/``HashFS.__len__()`` methods that return the count of all files.
 - Don't create the root directory (if it doesn't exist) until at least one file has been added.